// src/parser/types.ts
type PersistentBool = PersistentValue<boolean>;
type PersistentInt = PersistentValue<number>;
type GeoRock = PersistentValue<number>;

export type Scene<T extends number | boolean = boolean> =
  | (T extends boolean ? PersistentBool : never)
  | (T extends number ? PersistentInt : never);

type SceneData = {
  persistentBools: { serializedList: PersistentBool[] };
  persistentInts: { serializedList: PersistentInt[] };
  geoRocks: { serializedList: GeoRock[] };
};

export type SaveData = {
  playerData: PlayerData;
  sceneData: SceneData;
};